Crate yb_stats

yb_stats: a utility to extract all possible data from a YugabyteDB cluster.

This utility can:

Read YugabyteDB http endpoints and write the distinct groups of data into CSV files (--snapshot)

Read YugabyteDB http endpoints and report results back directory (--print-* without a snapshot number), and the adhoc snapshot mode.

Read yb_stats snapshots (CSV), and report the difference (--*-diff).

Read yb_stats snapshots (CSV), and report the snapshot data (--print-* <NR>).

This main file contains the Opts struct for commandline options via clap. It then calls the tasks using the opts structure.

Modules

clocks 🔒

Module for reading /tablet-server-clocks on the masters.

cluster_config 🔒

Module for reading /api/v1/cluster-config on the masters.

entities 🔒

Module for reading /dump-entities on the masters.

gflags 🔒

Module for reading /varz on the masters and tablet servers (deprecated: use the vars module).

health_check 🔒

Module for reading /api/v1/health-check on the masters.

isleader 🔒

Module for reading /api/v1/is-leader on the masters.

loglines 🔒

Module for reading /logs on the masters and tablet servers.

masters 🔒

Module for reading /api/v1/masters on the masters.

mems 🔒

Module for reading /memz on the masters and tablet servers.

memtrackers 🔒

Module for reading /mem-trackers on the masters and tablet servers.

metrics 🔒

Module for reading /metrics on the masters, tablet servers and YSQL webservers.

node_exporter 🔒

Module for reading /metrics on node exporter.

pprof 🔒

Module for reading /pprof/growth on the masters and tablet servers.

rpcs 🔒

Module for reading /rpcz on the masters, tablet servers and YSQL webservers.

snapshot 🔒

Module containing snapshot functions.

statements 🔒

Module for reading /statements from the YSQL webservers.

tablet_servers 🔒

Module for reading /api/v1/tablet-servers on the masters.

threads 🔒

Module for reading /threadz on the masters and tablet servers.

utility 🔒

Module for general utility functions.

vars 🔒

Module for reading /api/v1/varz from the masters and tablet servers.

versions 🔒

Module for reading /api/v1/version from the masters and tablet servers.

Structs

Opts

yb_stats switches

Constants

ACCEPT_INVALID_CERTS 🔒

Accept certificates not signed by an official CA?

DEFAULT_HOSTS 🔒

DEFAULT_PARALLEL 🔒

DEFAULT_PORTS 🔒

WRITE_DOTENV 🔒

Write the .env in the current working directory?

Functions

main 🔒

The entrypoint of the executable.