Module yb_stats::entities::functions

The impls and functions.

The /dump-entities endpoint contains a number of independent JSON arrays:

keyspaces: “keyspaces”: [{"keyspace_id":"00000000000000000000000000000001","keyspace_name":"system","keyspace_type":"ycql"},..]

tables: “tables”: [{"table_id":"000000010000300080000000000000af","keyspace_id":"00000001000030008000000000000000","table_name":"pg_user_mapping_user_server_index","state":"RUNNING"},..]

tablets: “tablets”: [{"table_id":"sys.catalog.uuid","tablet_id":"00000000000000000000000000000000","state":"RUNNING"},..]

3.1. replicas: “replicas”: [{"type":"VOTER","server_uuid":"047856aaf11547749694ca7d7941fb31","addr":"yb-2.local:9100"},..]

This is 3.1 because replicas are arrays nested in tablets.

The way these link together is:

tables.keyspace_id -> keyspaces.keyspace_id (keyspaces.keyspace_id must exist for tables.keyspace_id)

tables.table_id -> tablets.table_id, which contains the replicas as a nested array. tablets.table_id might not exist for tables.table_id, because some tables do not have tablets, such as the postgres catalog entries.

Special keyspaces:

keyspace	description

system	contains the local, partitions, roles, transactions, peers, size_estimates, transactions-<UUID> tables

system_schema	contains the YCQL indexes, views, aggregates, keyspaces, tables, types, functions, triggers, columns, sys.catalog tables

system_auth	contains the roles, role_permissions, resource_role_permissions_index tables

template1	postgres template1 database template, contains catalog

template0	postgres template0 database template, contains catalog

postgres	postgres standard database, not commonly used with YugabyteDB

yugabyte	postgres standard database, default database

system_platform	postgres database, contains a few extra catalog tables starting with ‘sql’

YCQL

YCQL requires a keyspace to be defined before user objects can be created and loaded, and keyspace, table and tablet will get a random UUID as id.

YSQL

YSQL keyspaces do get an id in a specific way. The id of the YSQL keyspace is in a format that later is used by the objects too. This is how a YSQL keyspace id looks like:

| 0  1  2  3| 4  6| 7| 8| 9|10 11|12 13 14 15|
|-----------|-----|--|--|--|-----|-----------|
|00 00 33 e5|00 00|30|00|80|00 00|00 00 00 00|
db                ver   var      object
oid               (3)   (8)      oid
(hex)                   (hex)

(position 12..15): A YSQL keyspace has the object id set to all 0. This is described in the YugabyteDB sourcecode: src/yb/common/entity_ids.cc

(position 7): Version 3 is for ISO4122 UUID version.

(position 9): Variant 8 means DCE 1.1, ISO/IEC 11578:1996

Version and variant are static currently.

YSQL object id and catalog data

The object OID number indicates whether an object is a catalog object or a user object. Any object OID lower than 16384 (0x4000) is a catalog object. This is why a user table_id always starts from ..4000.

YSQL colocated databases.

If a database is created with colocation turned on, it will generate a special table entry:

    {
      "table_id": "0000400f000030008000000000000000.colocated.parent.uuid",
      "keyspace_id": "0000400f000030008000000000000000",
      "table_name": "0000400f000030008000000000000000.colocated.parent.tablename",
      "state": "RUNNING"
    }

This indicates the keyspace/database is colocated, and thus any object not explicitly defined using its own tablets, will be stored in the tablets that are part of the database.

Functions

entity_diff

print_coloc_leader_host

print_entities