Module yb_stats::masters::structs

The structs

Structs

AppStatusPB

source: src/yb/common/wire_protocol.proto

CloudInfoPB

source: src/yb/common/common_net.proto

GetMasterRegistrationRequestPB

The main struct holding the master information.

HostPortPB

source: src/yb/common/common_net.proto

Masters

The root struct for deserializing /api/v1/masters.

MastersDiff

The wrapper struct for holding the btreemap holding the diff structs.

MastersDiffFields

The masters diff struct.

NodeInstancePB

source: src/yb/common/wire_protocol.proto

ServerRegistrationPB

source: src/yb/common/wire_protocol.proto

Enums

ErrorCode

source: src/yb/common/wire_protocol.proto

ErrorCodes

source: src/yb/common/wire_protocol.proto

PeerRole

source: src/yb/common/common_types.proto

Type Definitions

BTreeMastersDiff 🔒

BTreeMap for storing a master diff struct per permanent_uuid.