Module yb_stats::metrics::structs

The structs

Structs

AllMetricEntity

This struct is used by yb_stats as a wrapper for saving and loading the MetricEntity data.

Attributes

Struct to represent the attributes nested json found in a MetricEntity entity.

MetricDiffCountSum

The struct that holds the first and second snapshot statistics.

MetricDiffCountSumRows

The struct that holds the first and second snapshot statistics.

MetricDiffValues

The struct that holds the first and second snapshot statistics.

MetricEntity

The root struct for deserializing /metrics.

MetricEntityDiff

Wrapper struct for holding the diff btreemaps.

Enums

Metrics

Enum to represent the possible metric types found the nested metrics json structure. These can be a number of valid metric types:

Type Definitions

BTreeMetricDiffCountSum 🔒

BTreeMap for storing a metricentity countsum.

BTreeMetricDiffCountSumRows 🔒

BTreeMap for storing a metricentity countsum.

BTreeMetricDiffValues 🔒

BTreeMap for storing a metricentity value.