Module yb_stats::snapshot::functions

Snapshot functions

Functions

adhoc_diff

This function shows the difference report for the adhoc (in memory) snapshot data that allows to show a difference:

adhoc_metrics_diff

Take “adhoc” (in memory) snapshots of metrics only:

adhoc_nonmetrics_diff

perform_snapshot

The function to perform a general snapshot resulting in JSON files.

read_snapshot_json

This is the general yb_stat wide read_snapshot_json function.

save_snapshot_json

This is the general yb_stat wide save_snapshot_json function.

snapshot_diff

This function shows the difference report for the snapshot data that allows to show a difference:

snapshot_nonmetrics_diff

This function shows the difference report for the snapshot data that allows to show a difference: