Struct yb_stats::Opts

pub struct Opts {

    pub(crate) hosts: Option<String>,
    pub(crate) ports: Option<String>,
    pub(crate) parallel: Option<String>,
    pub(crate) stat_name_match: Option<String>,
    pub(crate) table_name_match: Option<String>,
    pub(crate) hostname_match: Option<String>,
    pub(crate) gauges_enable: bool,
    pub(crate) details_enable: bool,
    pub(crate) silent: bool,
    pub(crate) snapshot: bool,
    pub(crate) snapshot_comment: Option<String>,
    pub(crate) snapshot_diff: bool,
    pub(crate) snapshot_nonmetrics_diff: bool,
    pub(crate) entity_diff: bool,
    pub(crate) masters_diff: bool,
    pub(crate) versions_diff: bool,
    pub(crate) adhoc_metrics_diff: bool,
    pub(crate) adhoc_nonmetrics_diff: bool,
    pub(crate) snapshot_list: bool,
    pub(crate) begin: Option<i32>,
    pub(crate) end: Option<i32>,
    pub(crate) print_memtrackers: Option<Option<String>>,
    pub(crate) tail_log: bool,
    pub(crate) print_log: Option<Option<String>>,
    pub(crate) log_severity: String,
    pub(crate) print_entities: Option<Option<String>>,
    pub(crate) print_masters: Option<Option<String>>,
    pub(crate) print_tablet_servers: Option<Option<String>>,
    pub(crate) print_vars: Option<Option<String>>,
    pub(crate) print_version: Option<Option<String>>,
    pub(crate) print_rpcs: Option<Option<String>>,
    pub(crate) print_clocks: Option<Option<String>>,
    pub(crate) print_latencies: Option<Option<String>>,
    pub(crate) print_threads: Option<Option<String>>,
    pub(crate) print_gflags: Option<Option<String>>,
    pub(crate) print_cluster_config: Option<Option<String>>,
    pub(crate) print_health_check: Option<Option<String>>,
    pub(crate) disable_threads: bool,
    pub(crate) sql_length: usize,
    pub(crate) get_coloc_leader_host: Option<String>,

}

yb_stats switches

Fields
hosts: Option<String>

Snapshot input hostnames (comma separated)

ports: Option<String>

Snapshot input port numbers (comma separated)

parallel: Option<String>

Snapshot capture parallelism (default 1)

stat_name_match: Option<String>

Output filter for statistic names as regex

table_name_match: Option<String>

Output filter for table names as regex (requires –details-enable)

hostname_match: Option<String>

Output filter for hostname or ports as regex

gauges_enable: bool

Output setting to add statistics that are not counters

details_enable: bool

Output setting to increase detail, such as report each table and tablet individually

silent: bool

Snapshot setting to be as silent as possible, only errors are printed

snapshot: bool

Perform a snapshot (creates stored CSV files)

snapshot_comment: Option<String>

Snapshot add comment in snapshot overview

snapshot_diff: bool

Create a performance diff report using a begin and an end snapshot number.

snapshot_nonmetrics_diff: bool

Create a diff report using a begin and an end snapshot number without performance figures.

entity_diff: bool

Create an entity diff report using a begin and end snapshot number.

masters_diff: bool

Create a masters diff report using a begin and end snapshot number.

versions_diff: bool

Create a versions diff report using a begin and end snapshot number.

adhoc_metrics_diff: bool

Create an adhoc diff report only for metrics

adhoc_nonmetrics_diff: bool

Create an adhoc diff report excluding metrics

snapshot_list: bool

Lists the snapshots in the yb_stats.snapshots in the current directory.

begin: Option<i32>

Output setting to specify the begin snapshot number for diff report.

end: Option<i32>

Output setting to specify the end snapshot number for diff report.

print_memtrackers: Option<Option<String>>

Print memtrackers data for the given snapshot number

tail_log: bool

tail log data

print_log: Option<Option<String>>

Print log data for the given snapshot number

log_severity: String

Output log data severity to include: optional: I (use with –print_log)

print_entities: Option<Option<String>>

Print entity data for snapshot number, or get current.

print_masters: Option<Option<String>>

Print master server data for snapshot number, or get current.

print_tablet_servers: Option<Option<String>>

Print tablet server data for snapshot number, or get current.

print_vars: Option<Option<String>>

Print vars for snapshot number, or get current

print_version: Option<Option<String>>

Print version data for snapshot number, or get current.

print_rpcs: Option<Option<String>>

Print rpcs for the given snapshot number, or get current.

print_clocks: Option<Option<String>>

print clocks for the given snapshot number, or get current.

print_latencies: Option<Option<String>>

print master leader tablet server latencies

print_threads: Option<Option<String>>

Print threads data for the given snapshot number, or get current.

print_gflags: Option<Option<String>>

Print gflags for the given snapshot number, or get current.

print_cluster_config: Option<Option<String>>

Print cluster-config for the given snapshot number, or get current.

print_health_check: Option<Option<String>>

Print health-check for the given snapshot number, or get current.

disable_threads: bool

Snapshot disable gathering of thread stacks from /threadz

sql_length: usize

Output setting for the length of the SQL text to display

get_coloc_leader_host: Option<String>

Get the hostname for the tablet leader of a colocated YSQL database.