Struct yb_stats::tablet_servers::structs::TabletServer

pub struct TabletServer {

    pub hostname_port: Option<String>,
    pub timestamp: Option<DateTime<Local>>,
    pub tablet_server_hostname_port: Option<String>,
    pub time_since_hb: String,
    pub time_since_hb_sec: f32,
    pub status: String,
    pub uptime_seconds: u64,
    pub ram_used: String,
    pub ram_used_bytes: i64,
    pub num_sst_files: u64,
    pub total_sst_file_size: String,
    pub total_sst_file_size_bytes: i64,
    pub uncompressed_sst_file_size: String,
    pub uncompressed_sst_file_size_bytes: i64,
    pub path_metrics: Vec<PathMetrics>,
    pub read_ops_per_sec: f64,
    pub write_ops_per_sec: f64,
    pub disable_tablet_split_if_default_ttl: Option<bool>,
    pub user_tablets_total: i32,
    pub user_tablets_leaders: i32,
    pub system_tablets_total: i32,
    pub system_tablets_leaders: i32,
    pub active_tablets: i32,
    pub cloud: String,
    pub region: String,
    pub zone: String,

}

The main struct holding the tablet server information

This seems to have the majority of the information from src/yb/master/master_types.proto But have additional fields?

Fields
hostname_port: Option<String>

yb_stats added to allow understanding the source host.

timestamp: Option<DateTime<Local>>

yb_stats added to allow understanding the timestamp.

tablet_server_hostname_port: Option<String>

another added field to make this struct function in the deserializing using TabletServers and using it in the struct AllTabletServers.

time_since_hb: Stringtime_since_hb_sec: f32status: Stringuptime_seconds: u64ram_used: Stringram_used_bytes: i64num_sst_files: u64total_sst_file_size: Stringtotal_sst_file_size_bytes: i64uncompressed_sst_file_size: Stringuncompressed_sst_file_size_bytes: i64path_metrics: Vec<PathMetrics>read_ops_per_sec: f64write_ops_per_sec: f64disable_tablet_split_if_default_ttl: Option<bool>user_tablets_total: i32user_tablets_leaders: i32system_tablets_total: i32system_tablets_leaders: i32active_tablets: i32cloud: Stringregion: Stringzone: String